Add special lexer token __FreeBSD_version.
On FreeBSD, this adds the special token __FreeBSD_version which
corresponds to the result of the getosreldate() call on the target
system. This will make it possible to version the druntime bindings
based on the OS version, which is not currently possible in spite of the
fact that the actual C definitions sometimes changes from OS version to
OS version, and the result of getosreldate() is supposed to change every
time that the C bindings change on FreeBSD, so it gives us exactly what
we need to version the bindings.

This is particularly critical for the transition from FreeBSD 11 to
FreeBSD 12, since several, very important types related to file
operations had their definitions changed in incompatible ways, making it
so that it's not possible to use the same bindings for both OS versions,
which will make transitioning to supporting FreeBSD 12 difficult and
error-prone. While a binary built on FreeBSD 11 will run on 12 thanks to
extra effort the FreeBSD team went to to increase compatibility in spite
of the API and ABI changes, trying to build anything on FreeBSD 12 with
the current bindings doesn't work, and updating the bindings to matching
FreeBSD 12 would not only break building on FreeBSD 11, but the dance to
be able to update dmd and druntime to work on FreeBSD 12 without
versioning the bindings to work on both 11 and 12 is a pain. By far the
simplest solution is to fix it so that we can version the druntime
bindings so that they will work on both FreeBSD 11 and 12, and adding
__FreeBSD_version will allow us to do that - solving the problem not only
for FreeBSD 12 but for all future FreeBSD releases. It not only is the
easiest solution for those of us working on dmd and druntime, but it
allows users to use dmd on either FreeBSD 11 or 12 (assuming that it's
either built on FreeBSD 11 or on the target OS), whereas otherwise it
would only work on whichever version of the OS the druntime bindings
were for.

This change relates to https://issues.dlang.org/show_bug.cgi?id=17596,
though druntime will need to be updated with versioned bindings in a
separate PR to actually fix the bug. These changes simply make that
possible.

As to cross-compilation, in principle, a cross-compiler should provide a
way to override getosreldate when setting __FreeBSD_version to match the
target, otherwise the bindings could be wrong, but dmd isn't a
cross-compiler and I don't know how that would best fit into a
cross-compiler, so this commit makes no attempt to address that.